# docs/swagger.yaml
basePath: /api
definitions:
...
  admin.CreateListRequest:
    properties:
      description:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  admin.CreateSubscriberRequest:
    properties:
      email:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/domain.SubscriberStatus'
    type: object
  admin.CreateSubscribersRequest:
    properties:
      append:
        type: boolean
      subscribers:
        items:
          $ref: '#/definitions/admin.CreateSubscriberRequest'
        type: array
    type: object
...
  admin.DeleteResponse:
    properties:
      deleted:
        type: boolean
      message:
        type: string
    type: object
  admin.EmptyResponse:
    type: object
  admin.PaginatedListResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/admin.PaginationResponse'
    type: object
  admin.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  admin.UpdateListRequest:
    properties:
      description:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  admin.UpdateSubscriberRequest:
    properties:
      email:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/domain.SubscriberStatus'
    type: object
...
  domain.List:
    properties:
      created_at:
        description: Unix timestamp in seconds
        type: integer
      deleted_at:
        description: For soft deletes
        type: integer
      description:
        description: Description of the list
        type: string
      id:
        description: UUID
        type: string
      name:
        description: Name of the list
        type: string
      subscriber_count:
        description: Calculated field for subscriber count
        type: integer
      tags:
        description: Tags for categorization
        items:
          type: string
        type: array
      updated_at:
        description: Unix timestamp in seconds
        type: integer
    type: object
  domain.Subscriber:
    properties:
      created_at:
        description: Unix timestamp in seconds
        type: integer
      email:
        description: Unique email address
        type: string
      id:
        description: UUID
        type: string
      lists:
        items:
          $ref: '#/definitions/domain.SubscriberList'
        type: array
      name:
        description: Name of the subscriber
        type: string
      status:
        $ref: '#/definitions/domain.SubscriberStatus'
      updated_at:
        description: Unix timestamp in seconds
        type: integer
    type: object
  domain.SubscriberList:
    properties:
      created_at:
        description: Unix timestamp in seconds
        type: integer
      list_id:
        description: Foreign key to the List
        type: string
      status:
        $ref: '#/definitions/domain.SubscriberListStatus'
      subscribed_at:
        description: Unix timestamp of subscription
        type: integer
      unsubscribed_at:
        description: Unix timestamp of unsubscription
        type: integer
      updated_at:
        description: Unix timestamp in seconds
        type: integer
    type: object
  domain.SubscriberListStatus:
    enum:
    - confirmed
    - unsubscribed
    - bounced
    - complained
    type: string
    x-enum-varnames:
    - SubscriberListStatusConfirmed
    - SubscriberListStatusUnsubscribed
    - SubscriberListStatusBounced
    - SubscriberListStatusComplained
  domain.SubscriberStatus:
    enum:
    - enabled
    - disabled
    type: string
    x-enum-varnames:
    - SubscriberStatusEnabled
    - SubscriberStatusDisabled
info:
  contact:
    email: dev@headmail.io
    name: API Support
    url: https://headmail.io
  description: This is an email campaign API server.
  license:
    name: AGPL-3.0
    url: https://opensource.org/license/agpl-v3
  title: Headmail API
  version: "1.0"
paths:
...
  /lists:
    get:
      description: List all mailing lists
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Tags to filter by
        in: query
        items:
          type: string
        name: tags[]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.PaginatedListResponse'
      summary: List all mailing lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: Create a new mailing list
      parameters:
      - description: List to create
        in: body
        name: list
        required: true
        schema:
          $ref: '#/definitions/admin.CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.List'
      summary: Create a new mailing list
      tags:
      - lists
  /lists/{listID}:
    delete:
      description: Delete a mailing list
      parameters:
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.DeleteResponse'
      summary: Delete a mailing list
      tags:
      - lists
    get:
      description: Get a mailing list by ID
      parameters:
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.List'
      summary: Get a mailing list by ID
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: Update a mailing list
      parameters:
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: List to update
        in: body
        name: list
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.List'
      summary: Update a mailing list
      tags:
      - lists
  /lists/{listID}/subscribers:
    get:
      description: List subscribers
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.PaginatedListResponse'
      summary: List subscribers
      tags:
      - subscribers
    post:
      consumes:
      - application/json
      description: Add subscribers to a list
      parameters:
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: Subscribers to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreateSubscribersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin.EmptyResponse'
      summary: Add subscribers to a list
      tags:
      - subscribers
  /subscribers:
    get:
      description: List subscribers
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.PaginatedListResponse'
      summary: List subscribers
      tags:
      - subscribers
  /subscribers/{subscriberID}:
    delete:
      description: Delete a subscriber
      parameters:
      - description: Subscriber ID
        in: path
        name: subscriberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.DeleteResponse'
      summary: Delete a subscriber
      tags:
      - subscribers
    get:
      description: Get a subscriber by ID
      parameters:
      - description: Subscriber ID
        in: path
        name: subscriberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscriber'
      summary: Get a subscriber by ID
      tags:
      - subscribers
    put:
      consumes:
      - application/json
      description: Update a subscriber
      parameters:
      - description: Subscriber ID
        in: path
        name: subscriberID
        required: true
        type: string
      - description: Subscriber to update
        in: body
        name: subscriber
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateSubscriberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscriber'
      summary: Update a subscriber
      tags:
      - subscribers
...
swagger: "2.0"